feat: Add lab11 files
This commit adds the _ViewStart.cshtml and _ViewImports.cshtml files to the Views folder. The _ViewStart.cshtml file sets the default layout for views to _Layout.cshtml, while the _ViewImports.cshtml file imports the necessary namespaces and tag helpers for the views. These files improve the organization and functionality of the views in the application.